Fixing invoices

+++ database/migrations/2024_02_11_081944_create_entries_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('color')->nullable();
             $table->unsignedBigInteger('user_id');
             $table->integer('unit_price');
-            $table->double('vat');
+            $table->integer('vat');
             $table->boolean('left');
             $table->date('posting_date');
             $table->timestamps();
